ignore race condition
