+++ util/exec/exec_test.go
@@ -112,35 +112,6 @@ func TestRunCommand(t *testing.T) {
 	assert.Contains(t, entry.Data, "execID")
 }
 
-func TestRunCommandTimeout(t *testing.T) {
-	hook := test.NewGlobal()
-	log.SetLevel(log.DebugLevel)
-	defer log.SetLevel(log.InfoLevel)
-
-	output, err := RunCommand("sh", CmdOpts{Timeout: 500 * time.Millisecond}, "-c", "echo hello world && echo my-error >&2 && sleep 2")
-	assert.Equal(t, "hello world", output)
-	require.EqualError(t, err, "`sh -c echo hello world && echo my-error >&2 && sleep 2` failed timeout after 500ms")
-
-	assert.Len(t, hook.Entries, 3)
-
-	entry := hook.Entries[0]
-	assert.Equal(t, log.InfoLevel, entry.Level)
-	assert.Equal(t, "sh -c echo hello world && echo my-error >&2 && sleep 2", entry.Message)
-	assert.Contains(t, entry.Data, "dir")
-	assert.Contains(t, entry.Data, "execID")
-
-	entry = hook.Entries[1]
-	assert.Equal(t, log.DebugLevel, entry.Level)
-	assert.Equal(t, "hello world\n", entry.Message)
-	assert.Contains(t, entry.Data, "duration")
-	assert.Contains(t, entry.Data, "execID")
-
-	entry = hook.Entries[2]
-	assert.Equal(t, log.ErrorLevel, entry.Level)
-	assert.Equal(t, "`sh -c echo hello world && echo my-error >&2 && sleep 2` failed timeout after 500ms", entry.Message)
-	assert.Contains(t, entry.Data, "execID")
-}
-
 func TestRunCommandSignal(t *testing.T) {
 	hook := test.NewGlobal()
 	log.SetLevel(log.DebugLevel)
